i18n: Move tanslations to JSON files
