+++ .babelrc
@@ -1,6 +1,7 @@
 {
   "presets": ["@babel/preset-env", "@babel/preset-react"],
   "plugins": [
+    ["inline-json-import", {}],
     [
       "@babel/plugin-proposal-decorators",
       {
