+++ app/i18n/en/achievements.json
@@ -1,0 +1,49 @@
+{
+  "1": {
+    "title": "Welcome",
+    "description": "You've successfully created an account on CaptainFact"
+  },
+  "2": {
+    "title": "Not a robot",
+    "description": "With such a pretty email address, there's no doubt you're a human"
+  },
+  "3": {
+    "title": "Everybody hurts",
+    "description": "Looking for help in the help section was the good thing to do"
+  },
+  "4": {
+    "title": "Bulletproof",
+    "description": "With the browser extension activated, no one will fool you anymore"
+  },
+  "5": {
+    "title": "You're fake news!",
+    "description": "Well, seems you've destroyed this source"
+  },
+  "6": {
+    "title": "Social Addict",
+    "description": "You love having these US companies spying on your data, don't you?"
+  },
+  "7": {
+    "title": "Ambassador",
+    "description": "You are an ambassador of the community"
+  },
+  "8": {
+    "title": "Ghostbuster",
+    "description": "Thanks for reporting that bug!"
+  },
+  "9": {
+    "title": "Fact Check Me, I'm Famous",
+    "description": "You made your way up to the leaderboard, congrats Captain!"
+  },
+  "10": {
+    "title": "Artist",
+    "description": "One of your creations is integrated to the project"
+  },
+  "11": {
+    "title": "Good Vibes",
+    "description": "You made some suggestions for improvements or new features, and we find them interesting!"
+  },
+  "title": "Achievements",
+  "about": "About",
+  "unlocked": "You just unlocked the \"{{achievement}}\" achievement"
+}
