--- app/i18n/en/extension.js
@@ -1,8 +1,0 @@
-export default {
-  title: 'The browser extension',
-  subtitle: 'To fact-check YouTube',
-  description:
-    "The web extension injects CaptainFact's data directly on the content you're watching.",
-  moreInfo: 'It is 100% <1>open-source</1> and <3>respects your privacy</3>.',
-  addTo: 'Add to'
-}
