+++ app/i18n/en/extension.json
@@ -1,0 +1,7 @@
+{
+  "title": "The browser extension",
+  "subtitle": "To fact-check YouTube",
+  "description": "The web extension injects CaptainFact's data directly on the content you're watching.",
+  "moreInfo": "It is 100% <1>open-source</1> and <3>respects your privacy</3>.",
+  "addTo": "Add to"
+}
