+++ app/i18n/en/help.json
@@ -1,0 +1,30 @@
+{
+  "categories": {
+    "siteUsage": "Site usage",
+    "contribute": "Contribute",
+    "others": "Others"
+  },
+  "pages": {
+    "about": "About",
+    "achievements": "Achievements",
+    "ambassadors": "Become an ambassador",
+    "contributionGuidelines": "Contribution guidelines",
+    "credits": "Credits",
+    "jobs": "Join us!",
+    "moderation": "Moderation",
+    "privileges": "Privileges",
+    "reputation": "Reputation",
+    "bug_report": "Bug Report",
+    "extension": "Extension",
+    "privacy": "Privacy",
+    "contact": "Contact",
+    "censorship_requests": "Censorship Requests",
+    "contribute/code": "Programmers",
+    "contribute/tasks": "Independent tasks"
+  },
+  "title": "Help",
+  "goBack": "Go back to help index",
+  "aboutTitle": "About help pages",
+  "about": "The help pages are there to make it easier for you to get started with the platform. They are deliberately short and are strongly encouraged to be read before any participation to the fact-checking process.",
+  "toStart": "To get started, take a look at the <1>contribution guidelines</1> as well as explanations of <2>privileges</2> and the <3>reputation system</3>."
+}
