--- app/i18n/en/home.js
@@ -1,25 +1,0 @@
-export default {
-  presentationTitle: "Let's check the Internet",
-  presentation: `
-    Internet is filled with rich and abundant data. Information travels fast and we don't
-    have the time to check that everything we see and hear is correct.
-    <2></2>
-    <4></4>
-    <0>CaptainFact</0> is all about people debating and helping each other to
-    verify that data by sourcing it, voting and chatting about facts.  
-  `,
-  seeVideos: 'See all videos',
-  invitation: 'Ask for an invitation',
-  inviteFriend: 'Invite a friend',
-  footer: 'Created with <1/> using ',
-  emailPlaceholder: 'Email address',
-  inviteSuccess: 'Invitation request confirmed!',
-  registerAndFactCheck: 'Signup and fact-check',
-  installExtension: 'Install Extension',
-  partners: 'Our partners',
-  ambassadors: 'Ambassadors',
-  backers: 'Backers',
-  learnMore: 'Learn more',
-  aboutAmbassadors:
-    "Ambassadors are special members in charge of looking after the community. They also aim to report users' expectations to the technical team and participate in the governance of the project."
-}
