+++ app/i18n/en/home.json
@@ -1,0 +1,17 @@
+{
+  "presentationTitle": "Let's check the Internet",
+  "presentation": "\n    Internet is filled with rich and abundant data. Information travels fast and we don't\n    have the time to check that everything we see and hear is correct.\n    <2></2>\n    <4></4>\n    <0>CaptainFact</0> is all about people debating and helping each other to\n    verify that data by sourcing it, voting and chatting about facts.  \n  ",
+  "seeVideos": "See all videos",
+  "invitation": "Ask for an invitation",
+  "inviteFriend": "Invite a friend",
+  "footer": "Created with <1/> using ",
+  "emailPlaceholder": "Email address",
+  "inviteSuccess": "Invitation request confirmed!",
+  "registerAndFactCheck": "Signup and fact-check",
+  "installExtension": "Install Extension",
+  "partners": "Our partners",
+  "ambassadors": "Ambassadors",
+  "backers": "Backers",
+  "learnMore": "Learn more",
+  "aboutAmbassadors": "Ambassadors are special members in charge of looking after the community. They also aim to report users' expectations to the technical team and participate in the governance of the project."
+}
