+++ app/i18n/en/index.js
@@ -1,14 +1,14 @@
-import achievements from './achievements'
-import extension from './extension'
-import errors from './errors'
-import help from './help'
-import history from './history'
-import home from './home'
-import main from './main'
-import notifications from './notifications'
-import user from './user'
-import videoDebate from './videoDebate'
-import moderation from './moderation'
+import achievements from './achievements.json'
+import extension from './extension.json'
+import errors from './errors.json'
+import help from './help.json'
+import history from './history.json'
+import home from './home.json'
+import main from './main.json'
+import notifications from './notifications.json'
+import user from './user.json'
+import videoDebate from './videoDebate.json'
+import moderation from './moderation.json'
 
 export default {
   achievements,
