--- app/i18n/en/main.js
@@ -1,88 +1,0 @@
-export default {
-  entities: {
-    video: 'video',
-    video_plural: 'videos',
-    speaker: 'speaker',
-    speaker_plural: 'speakers',
-    statement: 'statement',
-    statement_plural: 'statements'
-  },
-  menu: {
-    login: 'Login',
-    signup: 'Sign up',
-    logout: 'Logout',
-    settings: 'Settings',
-    profile: 'Profile',
-    activity: 'Activity',
-    myActivity: 'My $t(menu.activity)',
-    content: 'Content',
-    language: 'Language',
-    other: 'Other',
-    donation: 'Support us',
-    extension: 'Download extension',
-    help: 'Learn more',
-    moderation: 'Moderation',
-    addedVideos: 'Added videos',
-    notifications: 'Notifications',
-    subscriptions: 'Subscriptions'
-  },
-  actions: {
-    save: 'Save',
-    cancel: 'Cancel',
-    send: 'Send',
-    moreInfo: 'More info',
-    goBack: 'Go back',
-    reload: 'Reload',
-    flag: 'Flag',
-    flagged: 'Flagged',
-    remove: 'Remove',
-    delete: 'Delete',
-    reply: 'Reply',
-    addToThread: 'Add to thread',
-    approve: 'Approve',
-    refute: 'Refute',
-    edit: 'Edit',
-    loading: 'Loading',
-    share: 'Share',
-    apply: 'Apply'
-  },
-  misc: {
-    timeAt: 'at',
-    staring: 'Staring',
-    by: 'By',
-    fieldMinLength: 'Must be at least {{min}} characters',
-    fieldMaxLength: 'Must be less than {{max}} characters',
-    and: 'and',
-    other: '{{count}} other',
-    other_plural: '{{count}} others',
-    character: 'character',
-    character_plural: 'characters',
-    clipboardSuccess: 'URL copied to clipboard',
-    clipboardFail: 'Cannot copy to clipboard',
-    all: 'All',
-    unknown: 'Unknown',
-    languageFilter: 'Language',
-    source: 'Source',
-    colon: ':',
-    flags: 'Flags'
-  },
-  videos: {
-    add: 'Add Video',
-    addThis: 'Add this video',
-    placeholder: 'Video URL',
-    analysing: 'Analysing video',
-    public: 'Public video',
-    unlisted: 'Unlisted',
-    publicDescription:
-      'Public videos are listed on videos page and in browser extension.',
-    unlistedDescription:
-      "Unlisted videos will not be listed on videos page nor in browsert extension. Use them to work on a content that you don't want to bring forward on the platform."
-  },
-  pagination: {
-    prev: 'Previous page',
-    next: 'Next page'
-  },
-  all: 'All',
-  partners: 'Partners',
-  users: 'Users'
-}
