--- app/i18n/en/moderation.js
@@ -1,23 +1,0 @@
-export default {
-  header: {
-    date: 'Date'
-  },
-  actions: {
-    flag_abusive: 'Refute flag',
-    unsure: 'Unsure',
-    confirm: 'Confirm flag'
-  },
-  reason: {
-    1: 'Personal attack or inappropriate language',
-    2: 'Unwanted commercial content or spam',
-    3: 'Irrelevant',
-    4: 'Not constructive'
-  },
-  seeContext: 'See in context',
-  whyReport: 'Why do you think this comment was reported?',
-  flaggedFor: 'flagged this for "$t(reason.{{reason}})"',
-  title: 'Moderation',
-  emptyModeration:
-    "There's nothing to moderate here. Go feed some trolls and come back later!",
-  pending: 'Pending moderation'
-}
