--- app/i18n/en/notifications.js
@@ -1,11 +1,0 @@
-export default {
-  empty: 'Your notifications box is empty',
-  seeAll: 'See all',
-  markAllAsRead: 'Mark all as seen',
-  markAllAsUnseen: 'Mark all as unseen',
-  filters: {
-    ALL: 'All',
-    SEEN: 'Seen',
-    UNSEEN: 'Unseen'
-  }
-}
