+++ app/i18n/en/notifications.json
@@ -1,0 +1,11 @@
+{
+  "empty": "Your notifications box is empty",
+  "seeAll": "See all",
+  "markAllAsRead": "Mark all as seen",
+  "markAllAsUnseen": "Mark all as unseen",
+  "filters": {
+    "ALL": "All",
+    "SEEN": "Seen",
+    "UNSEEN": "Unseen"
+  }
+}
