+++ app/i18n/en/user.json
@@ -1,0 +1,37 @@
+{
+  "username": "Username",
+  "realName": "Real name",
+  "password": "Password",
+  "passwordOptional": "Password - leave empty to keep it unchanged",
+  "repeatPassword": "Repeat password",
+  "email": "Email",
+  "emailOrUsername": "Email address or username",
+  "optional": "Optional",
+  "login": "Login",
+  "signup": "Sign up",
+  "needAnAccountQuestion": "Need an account?",
+  "alreadyHaveAccountQuestion": "Already have an account?",
+  "actionWithThirdParty": "Or login using a third party service:",
+  "deleteAccount": "Delete Account",
+  "deletedAccount": "Deleted Account",
+  "dangerZone": "Danger Zone",
+  "registeredSince": "Registered for <1/>",
+  "forgottenPassword": "Forgotten password",
+  "resetPassword": "Reset password",
+  "resetPasswordRequestSuccess": "If an account is associated with this address, an email with a link to reset your password has been sent",
+  "resetPasswordSuccess": "Password successfully changed",
+  "invitationOnlyTitle": "Ask for an invitation",
+  "invitationOnlyBody": "You must have an invitation to create an account. To request one, please enter you email below.",
+  "emailConfirmed": "Email confirmed",
+  "linkedAccounts": "Linked accounts",
+  "linkAccount": "Link account",
+  "unlinkAccount": "Unlink account",
+  "accountSettings": "Account Settings",
+  "newsletter": {
+    "unsubscribe_success": "You've been unsubscribed from our newsletter",
+    "unsubscribe_error": "Given token seems invalid. If the problem persist, contact us"
+  },
+  "noActivity": "No action recorded yet.",
+  "browsingAnalyzer": "Browsing analyzer",
+  "reputation": "Reputation"
+}
