+++ app/i18n/fr/achievements.json
@@ -1,0 +1,49 @@
+{
+  "1": {
+    "title": "Welcome",
+    "description": "Bienvenue dans l'aventure !"
+  },
+  "2": {
+    "title": "Pas un robot",
+    "description": "Avec une si belle adresse e-mail, vous ne pouvez pas être un robot"
+  },
+  "3": {
+    "title": "Everybody hurts",
+    "description": "Chercher de l'aide dans la section « Aide » était la bonne chose à faire"
+  },
+  "4": {
+    "title": "À l'épreuve des balles",
+    "description": "Avec l'extension pour navigateur installée, on ne vous la mettra plus à l'envers"
+  },
+  "5": {
+    "title": "You're fake news !",
+    "description": "On dirait que vous avec détruit cette source"
+  },
+  "6": {
+    "title": "Addict aux réseaux sociaux",
+    "description": "Vous aimez vous faire espionner par ces entreprises Américaines, n'est-ce pas ?"
+  },
+  "7": {
+    "title": "Ambassadeur",
+    "description": "Vous avez été nommé Ambassadeur de la communauté."
+  },
+  "8": {
+    "title": "Chasseur de fantômes",
+    "description": "Merci de nous avoir signalé ce bug !"
+  },
+  "9": {
+    "title": "Fact Check Me, I'm Famous",
+    "description": "Vous êtes arrivé en haut du classement ce mois-ci, bravo Capitaine !"
+  },
+  "10": {
+    "title": "Artiste",
+    "description": "Une ou plusieurs de vos créations sont intégrées dans le projet. Chapeau l'artiste !"
+  },
+  "11": {
+    "title": "Ondes Positives",
+    "description": "Une de vos suggestions d'améliorations ou de nouvelles fonctionnalités a été retenue."
+  },
+  "title": "Médailles",
+  "about": "À propos",
+  "unlocked": "Vous venez de déverrouiller le succès « {{achievement}} »"
+}
