--- app/i18n/fr/extension.js
@@ -1,8 +1,0 @@
-export default {
-  title: "L'extension navigateur",
-  subtitle: 'Pour vérifier les vidéos Youtube',
-  description:
-    "L'extension pour navigateur permet d'incruster les informations fournies sur CaptainFact directement sur le contenu que vous regardez.",
-  moreInfo: 'Elle est 100 % <1>open-source</1> et <3>respecte votre vie privée</3>.',
-  addTo: 'Ajouter à'
-}
