--- app/i18n/fr/help.js
@@ -1,32 +1,0 @@
-export default {
-  categories: {
-    siteUsage: 'Utilisation du site',
-    contribute: 'Contribuer',
-    others: 'Autres'
-  },
-  pages: {
-    about: 'À propos',
-    achievements: 'Médailles',
-    ambassadors: 'Devenir ambassadeur',
-    credits: 'Remerciements',
-    contributionGuidelines: 'Guide de contribution',
-    jobs: 'On recrute !',
-    moderation: 'Modération',
-    privileges: 'Privilèges',
-    reputation: 'Réputation',
-    bug_report: 'Signaler un bug',
-    extension: 'Extension',
-    privacy: 'Vie privée',
-    contact: 'Contact',
-    censorship_requests: 'Demandes de censure',
-    'contribute/code': 'Développeurs',
-    'contribute/tasks': 'Tâches indépendantes'
-  },
-  title: 'Aide',
-  goBack: "Retourner au menu d'aide principal",
-  aboutTitle: "À propos des pages d'aide",
-  about:
-    "Les pages d'aide sont là pour faciliter la prise en main de la plateforme. Elles sont volontairement courtes et leur lecture est fortement encouragée avant de participer au travail de vérification.",
-  toStart:
-    "Pour commencer, jetez un coup d'oeuil au <1>guide de contribution</1> ainsi qu'aux explications des <2>privilèges</2> et du <3>système de réputation</3>."
-}
