+++ app/i18n/fr/home.json
@@ -1,0 +1,17 @@
+{
+  "presentationTitle": "Contribuez à développer l'esprit critique et à vérifier les informations.",
+  "presentation": "\n    <0>CaptainFact</0> est un outil collaboratif de vérification des\n    vidéos YouTube.\n    <2></2>\n    Inscrivez-vous pour commencer à démentir, à confirmer ou à approfondir les\n    faits.\n    <4></4>\n    Installez notre extension pour bénéficier des données de CaptainFact directement\n    depuis votre lecteur YouTube.\n  ",
+  "seeVideos": "Voir toutes les vidéos",
+  "invitation": "Demander une invitation",
+  "inviteFriend": "Inviter un ami",
+  "footer": "Créé avec <1/> à l'aide d'",
+  "emailPlaceholder": "Adresse e-mail",
+  "inviteSuccess": "Demande d'invitation enregistrée !",
+  "registerAndFactCheck": "S'inscrire et vérifier",
+  "installExtension": "Installer l'extension",
+  "partners": "Les partenaires",
+  "ambassadors": "Ambassadeurs",
+  "backers": "Soutiens",
+  "learnMore": "En savoir plus",
+  "aboutAmbassadors": "Les ambassadeurs sont chargés de veiller sur la communauté. Ils ont aussi pour objectif de faire remonter les attentes des utilisateurs et de participer à la gouvernance du projet."
+}
