+++ app/i18n/fr/main.json
@@ -1,0 +1,86 @@
+{
+  "entities": {
+    "video": "vidéo",
+    "video_plural": "vidéos",
+    "speaker": "intervenant·e",
+    "speaker_plural": "intervenant·e·s",
+    "statement": "déclaration",
+    "statement_plural": "déclarations"
+  },
+  "menu": {
+    "login": "Se connecter",
+    "signup": "S'inscrire",
+    "logout": "Se déconnecter",
+    "settings": "Paramètres",
+    "profile": "Profil",
+    "activity": "Activité",
+    "myActivity": "Mon $t(menu.activity)",
+    "content": "Contenu",
+    "language": "Langue",
+    "other": "Autre",
+    "donation": "Nous soutenir",
+    "extension": "Télécharger l'extension",
+    "help": "Aide",
+    "moderation": "Modération",
+    "addedVideos": "Vidéos ajoutées",
+    "notifications": "Notifications",
+    "subscriptions": "Abonnements"
+  },
+  "actions": {
+    "save": "Sauvegarder",
+    "cancel": "Annuler",
+    "send": "Envoyer",
+    "moreInfo": "Plus d'informations",
+    "goBack": "Revenir en arrière",
+    "reload": "Actualiser",
+    "flag": "Signaler",
+    "flagged": "Signalé",
+    "remove": "Retirer",
+    "delete": "Supprimer",
+    "reply": "Répondre",
+    "addToThread": "Ajouter à la suite",
+    "approve": "Confirmer",
+    "refute": "Réfuter",
+    "edit": "Éditer",
+    "loading": "Chargement",
+    "share": "Partager",
+    "apply": "Appliquer"
+  },
+  "misc": {
+    "timeAt": "à",
+    "staring": "Avec",
+    "by": "Par",
+    "fieldMinLength": "Doit faire au moins {{min}} caractères",
+    "fieldMaxLength": "Doit faire moins de {{max}} caractères",
+    "and": "et",
+    "other": "{{count}} autre",
+    "other_plural": "{{count}} autres",
+    "character": "caractère",
+    "character_plural": "caractères",
+    "clipboardSuccess": "Adresse copiée dans le presse-papier",
+    "clipboardFail": "Impossible d'enregistrer l'adresse dans le presse-papier",
+    "all": "Tous",
+    "unknown": "Inconnu",
+    "languageFilter": "Langue",
+    "source": "Source",
+    "colon": " :",
+    "flags": "Signalements"
+  },
+  "videos": {
+    "add": "Ajouter une vidéo",
+    "addThis": "Ajouter cette vidéo",
+    "placeholder": "Adresse de la vidéo",
+    "analysing": "Analyse en cours de la vidéo",
+    "public": "Vidéo publique",
+    "unlisted": "Non listée",
+    "publicDescription": "Les vidéos publiques sont listées sur la page des vidéos et apparaissent dans l'extension pour navigateur.",
+    "unlistedDescription": "Les vidéos non listées n'apparaitront ni dans la liste des vidéos ni dans l'extension. Elles vous permettent de travailler sur un contenu sans que celui-ci ne soit mis en avant par la plateforme."
+  },
+  "pagination": {
+    "prev": "Page précédente",
+    "next": "Page suivante"
+  },
+  "all": "Toutes",
+  "partners": "Partenaires",
+  "users": "Utilisateurs"
+}
