--- app/i18n/fr/moderation.js
@@ -1,23 +1,0 @@
-export default {
-  header: {
-    date: 'Date'
-  },
-  actions: {
-    flag_abusive: 'Réfuter ce signalement',
-    unsure: 'Pas sûr',
-    confirm: 'Confirmer le signalement'
-  },
-  reason: {
-    1: 'Attaque personelle ou langage inapproprié',
-    2: 'Spam ou publicité non souhaitée',
-    3: 'Hors-sujet',
-    4: 'Non constructif'
-  },
-  seeContext: 'Voir dans son contexte',
-  whyReport: 'Selon vous, pourquoi ce commentaire a-t-il été signalé ?',
-  flaggedFor: 'a signalé ça pour "$t(reason.{{reason}})"',
-  title: 'Modération',
-  emptyModeration:
-    "Il n'y a rien à modérer ici. Allez nourrir quelques trolls et revenez plus tard.",
-  pending: 'En attente de moderation'
-}
