--- app/i18n/fr/notifications.js
@@ -1,27 +1,0 @@
-export default {
-  empty: "Vous n'avez aucune notification",
-  seeAll: 'Voir tout',
-  markAllAsRead: 'Tout marquer comme lu',
-  markAllAsUnseen: 'Tout marquer comme non lu',
-  filters: {
-    ALL: 'Toutes',
-    SEEN: 'Vues',
-    UNSEEN: 'Non vues'
-  },
-  message: {
-    newComment:
-      'Nouveau commentaire de <1>@{{username}}</1> sur <4>{{videoTitle}}</4> : « {{text}} »',
-    newSpeaker:
-      '<0>@{{username}}</0> a ajouté « {{speakerName}} » sur <6>{{videoTitle}}</6>',
-    removedStatement:
-      '<0>@{{username}}</0> a supprimé une citation sur <3>{{videoTitle}}</3>',
-    replyToComment:
-      '<0>@{{username}}</0> a répondu à votre commentaire sur <3>{{videoTitle}}</3> : « {{text}} »',
-    updatedVideo:
-      '<0>@{{username}}</0> a mis à jour les informations de la vidéo sur <3>{{videoTitle}}</3>',
-    updatedStatement:
-      '<0>@{{username}}</0> a mis à jour une citation sur <3>{{videoTitle}}</3> : "{{text}}"',
-    newStatement:
-      '<0>@{{username}}</0> a mis à jour une citation sur <3>{{videoTitle}}</3> : "{{text}}"'
-  }
-}
