+++ app/i18n/fr/user.json
@@ -1,0 +1,37 @@
+{
+  "username": "Nom d'utilisateur·rice",
+  "realName": "Votre vrai nom",
+  "password": "Mot de passe",
+  "passwordOptional": "Mot de passe - laisser vide pour conserver l'actuel",
+  "repeatPassword": "Répétez le mot de passe",
+  "email": "Adresse e-mail",
+  "emailOrUsername": "Adresse e-mail ou nom d'utilisateur",
+  "optional": "facultatif",
+  "login": "Se connecter",
+  "signup": "S'enregistrer",
+  "needAnAccountQuestion": "Besoin d'un compte ?",
+  "alreadyHaveAccountQuestion": "Vous avez déjà un compte ?",
+  "actionWithThirdParty": "Ou connectez-vous avec...",
+  "deleteAccount": "Supprimer le compte",
+  "deletedAccount": "Compte supprimé",
+  "dangerZone": "Zone dangereuse",
+  "registeredSince": "Inscrit·e depuis <1/>",
+  "forgottenPassword": "Mot de passe oublié",
+  "resetPassword": "Réinitialiser le mot de passe",
+  "resetPasswordRequestSuccess": "Si le compte existe, un e-mail avec un lien de réinitialisation vient d'être envoyé",
+  "resetPasswordSuccess": "Le mot de passe a bien été changé",
+  "invitationOnlyTitle": "Demander une invitation",
+  "invitationOnlyBody": "Vous devez avoir été invité pour créer un compte. Pour demander une invitation, entrez votre adresse e-mail ci-dessous.",
+  "emailConfirmed": "Adresse e-mail confirmée",
+  "linkedAccounts": "Comptes liés",
+  "linkAccount": "Lier le compte",
+  "unlinkAccount": "Délier le compte",
+  "accountSettings": "Paramètres du compte",
+  "newsletter": {
+    "unsubscribe_success": "Désinscription de la newsletter enregistrée !",
+    "unsubscribe_error": "Erreur lors de la desinscription, le code est probablement invalide. Si le problème persiste, contactez-nous"
+  },
+  "noActivity": "Aucune action n'a été effectuée pour le moment",
+  "browsingAnalyzer": "Analyse de la navigation",
+  "reputation": "Réputation"
+}
