--- app/i18n/fr/videoDebate.js
@@ -1,99 +1,0 @@
-export default {
-  debate: 'Vérification',
-  history: 'Historique',
-  approve: 'Confirme',
-  refute: 'Réfute',
-  comments: 'Commentaires',
-  community: 'Communauté',
-  flagForm: {
-    title: 'Pourquoi signaler ce commentaire ?',
-    xAvailable: '{{count}} disponnible',
-    xAvailable_plural: '{{count}} disponnibles'
-  },
-  comment: {
-    writeComment: 'Écrire un commentaire...',
-    addSource: 'Ajouter une source',
-    post: 'Ajouter en commentaire',
-    post_reply: 'Ajouter en réponse',
-    deleteThread: 'Supprimer le commentaire',
-    deleteThread_plural: 'Supprimer le fil de discussion',
-    replies_hide: 'Cacher les réponses',
-    replies_show: 'Montrer les réponses ({{count}})',
-    loadMore_replies: 'Charger plus de réponses ({{count}})',
-    loadMore_comments: 'Charger plus de commentaires ({{count}})',
-    replyingTo: 'En réponse à',
-    replyingToSelf: 'À la suite de',
-    approve: "J'approuve",
-    approve_reply: "J'approuve ce commentaire",
-    refute: 'Je réfute',
-    refute_reply: 'Je réfute ce commentaire',
-    revealForm: 'Ajouter une source ou un commentaire',
-    invalidURL: "Le lien vers la source n'est pas valide"
-  },
-  video: {
-    addedBy: 'Ajoutée par un {{userType}}',
-    partner: 'partenaire',
-    user: 'utilisateur',
-    edit: 'Editer la vidéo',
-    shiftStatements: 'Décaler toutes les affirmations (en secondes)'
-  },
-  speaker: {
-    one: 'Intervenant·e',
-    remove: "Retirer l'intervenant·e",
-    confirmRemove:
-      'Toutes les déclarations de {{speaker.full_name}} sur cette vidéo vont devenir orphelines',
-    add: 'Ajouter un·e intervenant·e',
-    create: 'Créer l\'intervenant·e "{{name}}"',
-    edit: 'Modifier les informations de "{{name}}"',
-    search: 'Tapez au moins 3 lettres pour lancer la recherche',
-    fullName: 'Nom complet',
-    title: 'Titre',
-    titlePlaceholder: 'Politicien, docteur, journaliste...',
-    noneFound: 'Aucun intervenant trouvé pour ce nom'
-  },
-  statement: {
-    remove: 'Retirer la citation',
-    confirmRemove: 'Êtes-vous sûr·e ?',
-    textPlaceholder: "Retranscrivez ici les propos de l'intervenant·e",
-    noSpeakerTextPlaceholder:
-      "Décrivez ce qui apparait à l'image ou ajoutez un intervenant",
-    text: 'Le texte',
-    add: 'Ajouter une citation',
-    abortAdd: "Annuler l'ajout de citation",
-    autoscroll_enable: 'Activer le défilement automatique',
-    autoscroll_disable: 'Désactiver le défilement automatique',
-    soundOnBackgroundFocus_enable:
-      'Jouer un son quand la citation change en arriere plan',
-    soundOnBackgroundFocus_disable:
-      'Ne pas jouer de son quand la citation change en arriere plan',
-    reverseTimeLock_lock: 'Déverrouiller la position',
-    reverseTimeLock_unlock: 'Verrouiller la position'
-  },
-  loading: {
-    statements: 'Chargement des textes',
-    video: 'Chargement de la video'
-  },
-  tips: {
-    firstStatement:
-      "Vous pouvez <1>ajouter une citation</1> en cliquant sur l'icone <3></3> à côté du nom de l'intervenant",
-    firstSpeaker:
-      'Commencez par ajouter un intervenant en utilisant la colonne de gauche',
-    noContentUnauthenticated:
-      "Cette vidéo n'a pas encore été vérifiée. Connectez-vous pour contribuer !"
-  },
-  presence: {
-    viewer: '{{count}} spectateur',
-    viewer_plural: '{{count}} spectateurs',
-    user: '{{count}} fact-checker',
-    user_plural: '{{count}} fact-checkers'
-  },
-  warningUnlisted:
-    "Cette vidéo n'est pas listée publiquement. Elle n'a peut-être pas reçu la visibilité nécessaire pour que les sources et citations soient correctement discutées.",
-  wikidata: {
-    id: 'Numéro unique Wikidata (QID)',
-    search: 'Tapez un nom pour lancer la recherche sur Wikipédia',
-    autofill: 'Récupérer automatiquement depuis Wikipédia',
-    using: 'Données extraites de <2>{{qid}}</2> (<5>éditer</5>)',
-    suggestions: 'Suggestions Wikidata :'
-  }
-}
