+++ package.json
@@ -130,6 +130,7 @@
     "babel-eslint": "^10.0.1",
     "babel-jest": "^24.7.1",
     "babel-loader": "^8.0.0",
+    "babel-plugin-inline-json-import": "^0.3.2",
     "babel-plugin-react-remove-properties": "^0.3.0",
     "babel-plugin-styled-components": "^1.9.2",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
